Criacao main_view.dart, criacao folder repository dentro da home
