+++ lib/features/home/repository/categories_list.dart
@@ -1,0 +1,18 @@
+import 'package:my_app/features/home/repository/category.dart';
+
+final List<Category> categoriesList = [
+  Category(
+    name: 'Comidas',
+    url:
+        'https://www.istoedinheiro.com.br/wp-content/uploads/sites/17/2019/08/din1135-sustenta5.jpg',
+  ),
+  Category(
+    name: 'Lanches',
+    url:
+        'https://img.freepik.com/fotos-premium/classico-cheeseburger-com-carne-de-queijo-bacon-tomate-cebola-e-alface-isolado-no-fundo-branco_183587-963.jpg',
+  ),
+  Category(
+    name: 'Pizza',
+    url: 'https://www.pngmart.com/files/1/Pepperoni-Pizza.png',
+  ),
+];
