+++ lib/features/home/repository/promotion_list.dart
@@ -1,0 +1,29 @@
+import 'package:my_app/features/home/repository/promotion.dart';
+
+final promotionList = [
+  Promotion(
+    name: 'Promo 1',
+    url:
+        'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSykY9ZDB3QYHKVYoc5q7yXsFKlT7vtYrs-Mg&usqp=CAU',
+  ),
+  Promotion(
+    name: 'Promo 2',
+    url:
+        'https://static-images.ifood.com.br/image/upload/t_medium/pratos/a3e32426-7659-484b-9309-bf3249bcaf62/202106301645_1VSY_i.jpg',
+  ),
+  Promotion(
+    name: 'Promo 3',
+    url:
+        'https://blog.consumer.com.br/wp-content/uploads/2020/02/ifood-loop.jpg',
+  ),
+  Promotion(
+    name: 'Promo 4',
+    url:
+        'https://i1.wp.com/surgiu.com.br/wp-content/uploads/2018/08/Mc-iFood-mcdiafeliz180824_121012.jpg?resize=1280%2C720',
+  ),
+  Promotion(
+    name: 'Promo 5',
+    url:
+        'https://gkpb.com.br/wp-content/uploads/2019/08/cropped-ifood-oferecera-pratos-de-R-1-para-novos-usuarios-geek-publicitario.jpg.png',
+  ),
+];
