+++ lib/features/home/view/widgets/category_view.dart
@@ -1,0 +1,70 @@
+import 'package:flutter/material.dart';
+import 'package:my_app/features/home/repository/category.dart';
+
+class CategoriesView extends StatelessWidget {
+  const CategoriesView({
+    super.key,
+    required this.categoryList,
+  });
+
+  final List<Category> categoryList;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.all(24),
+      child: SizedBox(
+        height: 150,
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: <Widget>[
+            const Text(
+              'Categorias',
+              style: TextStyle(fontSize: 18),
+            ),
+            Expanded(
+              child: ListView.builder(
+                shrinkWrap: true,
+                scrollDirection: Axis.horizontal,
+                itemCount: categoryList.length,
+                itemBuilder: (BuildContext context, int index2) => Card(
+                  child: Column(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      SizedBox(
+                        height: 80,
+                        width: 150,
+                        child: Image.network(
+                          categoryList[index2].url,
+                          errorBuilder: (context, error, stackTrace) {
+                            return Container(
+                              color: Colors.amber,
+                              alignment: Alignment.center,
+                              child: const Text(
+                                'Whoops!',
+                                style: TextStyle(fontSize: 30),
+                              ),
+                            );
+                          },
+                        ),
+                      ),
+                      const SizedBox(height: 16),
+                      Text(
+                        categoryList[index2].name,
+                        style: const TextStyle(
+                          fontSize: 14,
+                          color: Color.fromARGB(255, 71, 75, 80),
+                        ),
+                      ),
+                    ],
+                  ),
+                ),
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
