+++ lib/features/home/view/widgets/gourmet_widget.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+
+class GourmetRestaurants extends StatelessWidget {
+  const GourmetRestaurants({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.only(top: 10),
+      padding: const EdgeInsets.all(12),
+      child: ClipRRect(
+        borderRadius: BorderRadius.circular(4),
+        child: Image.network(
+          'https://www.istoedinheiro.com.br/wp-content/uploads/sites/17/2019/08/din1135-sustenta5.jpg',
+          errorBuilder: (context, error, stackTrace) {
+            return Container(
+              color: Colors.amber,
+              alignment: Alignment.center,
+              child: const Text(
+                'Whoops!',
+                style: TextStyle(fontSize: 30),
+              ),
+            );
+          },
+        ),
+      ),
+    );
+  }
+}
