+++ pubspec.yaml
@@ -25,6 +25,3 @@ dev_dependencies:
 flutter:
   uses-material-design: true
   generate: true
-
-  assets:
-    - assets/
